Add deployAndGenerateInvite CLI to avoid running server
