+++ packages/server/package.json
@@ -6,6 +6,7 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "ts-node index.ts",
+    "deployAndGenerateInvite": "ts-node deploy-and-generate-invite.ts",
     "rootGen": "npx ts-node rootInviteGen.ts",
     "generateClient": "open-rpc-generator generate -c open-rpc-generator-config.json"
   },
